+++ changelog.rst
@@ -8,3 +8,4 @@ Version 1.2
 Release date: Not released
 
 - Added environment based config overriding
+- Added json response to View handler